css/css-will-change/will-change-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>will-change is not animatable</title>
<link rel="help" href="https://drafts.csswg.org/css-will-change-1/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="target"></div>
<script>
/*
 * The `will-change` property must not be animatable.
 */
promise_test(async (t) => {
  const target = document.getElementById('target');
  t.add_cleanup(() => {
    target.style.transition = '';
    target.style.willChange = '';
  });

  // 1. Apply a CSS transition for 'will-change' on the target element.
  target.style.willChange = 'auto';
  target.style.transition = 'will-change 1s';

  // Force layout to ensure the initial state is computed
  getComputedStyle(target).willChange;

  target.addEventListener('transitionrun', t.unreached_func('transitionrun should not be fired'));
  target.addEventListener('transitionstart', t.unreached_func('transitionstart should not be fired'));

  // 2. Change the 'will-change' property from 'auto' to 'transform'.
  target.style.willChange = 'transform';

  // 3. Check the computed style of 'will-change' immediately after the change.
  assert_equals(getComputedStyle(target).willChange, 'transform', 'The computed value should be "transform" immediately.');

  // Wait a couple of frames to ensure no transition events are dispatched
  await new Promise(resolve => requestAnimationFrame(() => requestAnimationFrame(resolve)));
}, "will-change is not animatable");
</script>
</body>
